% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_distance}
\alias{calc_distance}
\title{두 좌표의 거리 구하기}
\usage{
calc_distance(
  lon1,
  lat1,
  lon2,
  lat2,
  proj = c("WGS84", "Bessel", "GRS80", "KATECH")
)
}
\arguments{
\item{lon1}{numeric. 첫 좌표의 경도.}

\item{lat1}{numeric. 첫 좌표의 위도.}

\item{lon2}{numeric. 둘째 좌표의 경도.}

\item{lat2}{numeric. 둘째 좌표의 위도.}

\item{proj}{character. 좌표계 CRS 정보.}
}
\value{
numeric. 두 점 간의 거리. 미터(m) 단위의 거리.
}
\description{
좌표계에서 위도, 경도로 이루어진 두 지점간의 거리를 구함
}
\details{
proj에 사용할 수 있는 좌표계 CRS정보는 다음과 같습니다. 
만약 다음 4개가 아닌 좌표계는 CRS 문자열을 기술해야 합니다.:
\itemize{
\item "WGS84" : WGS84 경위도 좌표계. (EPSG:4326)
\item "Bessel" : Bessel 1841 경위도. 한국과 일본에 잘 맞는 지역타원체를 사용한 좌표계.
\item "GRS80" : 통계청 통계지리정보서비스 좌표계(네이버지도에서 사용중인 좌표계). (EPSG:5179)
\item "KATECH" : KATECH 좌표계. 비표준 좌표계임.
}
}
\examples{
calc_distance(37.23, 132.12, 37.32, 133.45)

}
