% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_projection}
\alias{convert_projection}
\title{경위도 좌표계 위치정보의 좌표계 변환}
\usage{
convert_projection(
  x,
  y,
  from = c("WGS84", "Bessel", "GRS80", "KATECH", "ESPG:5181")[3],
  to = c("WGS84", "Bessel", "GRS80", "KATECH", "ESPG:5181")
)
}
\arguments{
\item{x}{numeric. 변환할 위치 경도}

\item{y}{numeric. 변환할 위치 위도}

\item{from}{character. 변경 전의 좌표계 CRS 정보}

\item{to}{character. 변경 후의 좌표계 CRS 정보}
}
\value{
data.frame. 변경 전후의 경도, 위도 정보.
데이터 프레임의 변수는 다음과 같습니다.
\itemize{
\item x numeric. 변경 전의 경도
\item y numeric. 변경 전의 위도
\item lon numeric. 변경 후의 경도
\item lat numeric. 변경 후의 위도
}
}
\description{
경위도 좌표계 위치정보인 (경도, 위도) 위치 정보를 다른 좌표계의 값으로 변환
}
\details{
from과 to에 사용할 수 있는 좌표계 CRS정보는 다음과 같습니다. 
만약 다음 4개가 아닌 좌표계는 CRS 문자열을 기술해야 합니다.:
\itemize{
\item "WGS84" : WGS84 경위도 좌표계. (EPSG:4326)
\item "Bessel" : Bessel 1841 경위도. 한국과 일본에 잘 맞는 지역타원체를 사용한 좌표계.
\item "GRS80" : 통계청 통계지리정보서비스 좌표계(네이버지도에서 사용중인 좌표계). (EPSG:5179)
\item "ESPG:5181" : 국토지리원 옛 표준. 중부원점 좌표계(Kakao, 공공데이터포탈에서 사용중인 좌표계). (ESPG:5181)
\item "KATECH" : KATECH 좌표계. 비표준 좌표계임.
}
}
\examples{
\donttest{
x <- c(126.9691, 127.4926) 
y <- c(37.56825, 36.23795)

convert_projection(x, y, from = "WGS84", to = "GRS80")

x <- c(953116.3, 999335.1) 
y <- c(1952231, 1804526) 
convert_projection(x, y, from = "GRS80", to = "WGS84")
}
}
