% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polygons.R
\docType{data}
\name{mega}
\alias{mega}
\alias{cty}
\alias{admi}
\title{광역시도, 시군구, 읍면동 수치지도 및 통계}
\format{
기준일자, 6개의 행정구역 코드 및 값과 개별 통계정보를 담은 sf 클래스 객체.
\describe{
  \item{BASE_DATE}{character. 기준일자.}
  \item{MEGA_CD}{character. 광역시도 코드.}
  \item{MEGA_NM}{character. 광역시도 이름.}
  \item{CTY_CD}{character. 시군구 코드.}
  \item{CTY_NM}{character. 시군구 이름.}
  \item{ADMI_CD}{character. 읍면동 코드.}
  \item{ADMI_NM}{character. 읍면동 이름.}
  \item{LAND_AREA}{numeric. 면적(km^2).}
  \item{geometry}{MULTIPOLYGON. 지도 polygons.}
}
}
\source{
"통계청 통계지리정보서비스" in <https://sgis.kostat.go.kr>, License : 공공저작물 자유이용허락 표시기준(공공누리, KOGL) 제 1유형
}
\usage{
data(mega)

data(cty)

data(admi)
}
\description{
sf 클래스 객체로 만들어진 광역시도(mega), 시군구(cty), 읍면동(admi) 레벨의 수치지도 및 관련 통계
}
\details{
sf 클래스 객체로 만들어진 데이터로서 2022년 6월말 기준의 데이터입니다. 
읍면동 레벨의 데이터는 상위 시군구, 광역시도 정보를 포함하며, 
시군구 레벨의 데이터는 상위 광역시도 정보를 포함합니다.
}
\keyword{datasets}
