% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stats.R
\docType{data}
\name{mega_population_age}
\alias{mega_population_age}
\alias{cty_population_age}
\alias{admi_population_age}
\title{광역시도, 시군구, 읍면동 성별 연령대별 인구수}
\format{
기준일자, 6개의 행정구역 코드 및 값과 개별 통계정보를 담은 tibble 클래스 객체.
\describe{
  \item{base_ym}{character. 기준월도.}
  \item{mega_cd}{character. 광역시도 코드.}
  \item{mega_nm}{character. 광역시도 이름.}
  \item{cty_cd}{character. 시군구 코드.}
  \item{cty_nm}{character. 시군구 이름.}
  \item{admi_cd}{character. 읍면동 코드.}
  \item{admi_nm}{character. 읍면동 이름.}
  \item{age_group}{character. 5세 단위의 연령그룹.}
  \item{gender}{character. 성별.}
  \item{population}{numeric. 인구수.}
}
}
\source{
"주민등록 인구통계" in <https://jumin.mois.go.kr/index.jsp>.
}
\usage{
data(mega_population_age)

data(cty_population_age)

data(admi_population_age)
}
\description{
광역시도(mega), 시군구(cty), 읍면동(admi) 레벨의 성별 연령대별 인구 통계
}
\details{
tibble 클래스 객체로 만들어진 데이터로서 2022년 6월 기준의 데이터입니다. 
읍면동 레벨의 데이터는 상위 시군구, 광역시도 정보를 포함하며, 
시군구 레벨의 데이터는 상위 광역시도 정보를 포함합니다.
}
\keyword{datasets}
