% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimal_map_size}
\alias{optimal_map_size}
\title{최적 지도 이미지 사이즈 계산}
\usage{
optimal_map_size(map, width = 800, height = NULL)
}
\arguments{
\item{map}{sf 혹은 sp 객체. 계산한 지도 객체.}

\item{width}{numeric. 너비의 크기. 기본값은 800.}

\item{height}{numeric. 높이의 크기.}
}
\value{
list. 높이와 너비의 크기.
리스트의 성분은 다음과 같습니다.
\itemize{
\item width : 너비의 크기.
\item height : 높이의 크기.
}
}
\description{
플롯팅할 지도의 종횡비를 고려하여, 최적의 지도 이미지 크기를 계산
}
\details{
sp객체나 sf 객체의 수치지도의 종횡비를 계산한 후, 지정한 너비(경도)나 높이(위도) 
에 대응하는 최적의 높이나 너비의 크기를 계산합니다.
}
\examples{
\donttest{
optimal_map_size(mega)
optimal_map_size(mega, height = 600)
optimal_map_size(mega \%>\% filter(mega_nm \%in\% "서울특별시"))
}
}
