% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{position2mega}
\alias{position2mega}
\alias{position2cty}
\alias{position2admi}
\title{위도/경도로 행정구역 코드와 이름 가져오기}
\usage{
position2mega(x, y, proj = c("WGS84", "Bessel", "GRS80", "KATECH"))

position2cty(x, y, proj = c("WGS84", "Bessel", "GRS80", "KATECH"))

position2admi(x, y, proj = c("WGS84", "Bessel", "GRS80", "KATECH"))
}
\arguments{
\item{x}{numeric. 행정구역 정보를 얻을 경도}

\item{y}{numeric. 행정구역 정보를 얻을 위도}

\item{proj}{character. 위치정보의 좌표계 CRS 정보}
}
\value{
data.frame. 행정구역 정보.
데이터 프레임의 변수는 다음과 같습니다.
\itemize{
\item lon numeric. 경도.
\item lat numeric. 위도.
\item base_ym character. 경계 수치지도의 기준 년월.
\item mega_cd character. 광역시도 코드.
\item mega_nm character. 광역시도 이름.
\item cty_cd character. 시군구 코드.
\item cty_nm character. 시군구 이름.
\item admi_cd character. 읍면동 코드.
\item admi_nm character. 읍면동 이름.
}
}
\description{
경위도 좌표계 위치정보인 (경도, 위도) 위치 정보로 행정구역 정보를 구함
}
\details{
proj에 사용할 수 있는 좌표계 CRS정보는 다음과 같습니다. 
만약 다음 4개가 아닌 좌표계는 CRS 문자열을 기술해야 합니다.:
\itemize{
\item "WGS84" : WGS84 경위도 좌표계. (EPSG:4326)
\item "Bessel" : Bessel 1841 경위도. 한국과 일본에 잘 맞는 지역타원체를 사용한 좌표계.
\item "GRS80" : 통계청 통계지리정보서비스 좌표계(네이버지도에서 사용중인 좌표계). (EPSG:5179)
\item "KATECH" : KATECH 좌표계. 비표준 좌표계임.
}
}
\examples{
\donttest{
x <- c(126.9691, 127.4926, 157.4926) 
y <- c(37.56825, 36.23795, 36.23795)

# 광역시도 정보 구하기
position2mega(x, y)

# 시군구 정보 구하기
position2cty(x, y)

# 읍면동 정보 구하기
position2admi(x, y)

x <- c(953116.3, 999335.1, 3728768.6) 
y <- c(1952231, 1804526, 2250492)

position2mega(x, y, proj = "GRS80")
}
}
