% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stats.R
\docType{data}
\name{post_admi}
\alias{post_admi}
\title{우편번호 행정동 매핑 데이터}
\format{
8개의 변수, 29,241건을 담은 tbl_df 클래스 객체.
\describe{
  \item{post_cd}{character. 도로명 주소 우편번호}
  \item{base_ym}{character. 기준월도.}
  \item{mega_cd}{character. 광역시도 코드.}
  \item{mega_nm}{character. 광역시도 이름.}
  \item{cty_cd}{character. 시군구 코드.}
  \item{cty_nm}{character. 시군구 이름.}
  \item{admi_cd}{character. 읍면동 코드.}
  \item{admi_nm}{character. 읍면동 이름.}
}
}
\source{
"주소기반산업지원서비스 사이트의 도로명주소 한글" in <https://business.juso.go.kr/addrlink/attrbDBDwld/attrbDBDwldList.do?cPath=99MD&menu=도로명주소+한글>.
}
\usage{
data(post_admi)
}
\description{
우편번호로 광역시도 > 시군구 > 읍면동을 매핑하기 위한 메타 데이터
}
\details{
이 데이터셋은 우편번호 데이터로 읍면동 레벨까지의 행정구역을 매핑하기 위해서 
제작된 데이터셋으로, 모든 우편번호를 포함하지 않음. 그 이유는 우편번호만으로 
행정동 기준은 읍면동 레벨까지 매핑시킬 수 없는 구조이기 때문임. 
그래서 우편번호로 광역시도 > 시군구 > 읍면동을 매핑할 수 있는 사례만 발췌한 것임.
}
\keyword{datasets}
