% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stats.R
\docType{data}
\name{school}
\alias{school}
\title{전국 초중등학교 위치 표준 데이터}
\format{
기준일자, 6개의 행정구역 코드 및 값과 개별 통계정보를 담은 tibble 클래스 객체.
\describe{
  \item{school_id}{character. 학교 ID.}
  \item{school_nm}{character. 학교 이름.}
  \item{school_class}{character. 학교 급구분.} 
  \item{estab_date}{character. 설립 일자.} 
  \item{estab_class}{character. 설립 형태.} 
  \item{branch_cd}{character. 본교분교 구분.} 
  \item{operate_stat}{character. 운영 상태.} 
  \item{addr_land}{character. 소재지 지번 주소.} 
  \item{addr_road}{character. 소재지 도로명 주소.} 
  \item{edu_office_cd}{character. 시도 교육청 코드.} 
  \item{edu_office_nm}{character. 시도 교육청 이름.} 
  \item{edu_support_cd}{character. 교육지원청 코드.} 
  \item{edu_support_nm}{character. 교육지원청 이름.} 
  \item{lat}{numeric. 위도.}
  \item{lon}{numeric. 경도.}
  \item{base_date}{character. 기준일자.}
  \item{mega_cd}{character. 광역시도 코드.}
  \item{mega_nm}{character. 광역시도 이름.}
  \item{cty_cd}{character. 시군구 코드.}
  \item{cty_nm}{character. 시군구 이름.}
  \item{admi_cd}{character. 읍면동 코드.}
  \item{admi_nm}{character. 읍면동 이름.}
}
}
\source{
공공데이터포털의 "전국초중등학교위치표준데이터" in <https://www.data.go.kr/data/15021148/standard.do?recommendDataYn=Y>.
}
\usage{
data(school)
}
\description{
초중등교육법에 따라 설립 승인을 받아 운영하는 초등학교, 중학교, 고등학교의 소재지 정보
}
\details{
tibble 클래스 객체로 만들어진 데이터로서 2022년 9월 21일 기준의 데이터입니다. 
학교 위치정보와 속성, 행정구역 코드 매핑 정보 등이 포함되어 있습니다.
}
\keyword{datasets}
