% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{thematic_map}
\alias{thematic_map}
\title{주제도 시각화}
\usage{
thematic_map(
  zoom = c("mega", "cty", "admi")[1],
  subset = NULL,
  stat = NULL,
  polygon = TRUE,
  point = FALSE,
  label = NULL,
  col_cnt = 9,
  palette = "YlOrRd",
  line_col = "darkgray",
  fill = "lightblue",
  point_col = "blue",
  title = NULL,
  subtitle = NULL,
  legend_pos = c("none", "right", "left", "bottom", "top"),
  base_family = "NanumSquare"
)
}
\arguments{
\item{zoom}{character. "mega", "cty", "admi"는 각각 광역시도,
시군구, 읍면동 레벨의 주제도를 지정함. 기본값은 "mega"로 광역시도 지도 정보를 가져옴.}

\item{subset}{expression. 지도에서 일부 영역을 표현할 조건.}

\item{stat}{character. 주제도에서 표현할 통계의 종류 코드.}

\item{polygon}{logical. 다각형 영역에 heatmap을 표현할지의 여부. 
기본값은 TRUE. stat 인수를 사용할 경우에만 허용됨.}

\item{point}{logical. 다각형 영역에 산점도 표현할지의 여부. 
기본값은 FALSE. stat 인수를 사용할 경우에만 허용됨.}

\item{label}{character. 다각형 영역에 범례 라벨을 출력하는 방법을 설정함.
기본값은 NULL로 라벨을 출력하지 않음. "name"은 행정구역의 이름을, 
"value"는 통계값을 라벨링함. "all"은 행정구역 이름과 통계값을 라벨링함.}

\item{col_cnt}{numeric. 주제도에서 표현할 색상 크라데이션의 개수.}

\item{palette}{character. 주제도에서 표현할 색상 팔레트 이름.}

\item{line_col}{character. 지도의 경계를 채울 색상을 지정함. 기본값은 "darkgray".}

\item{fill}{character. 지도의 다각형 내부를 채울 색상을 지정함. stat 인수의
값을 지정하지 않을 경우에만 적용됨. 기본값은 "lightblue".}

\item{point_col}{character. 통계의 크기를 표현할 포인트를 채울 색상을 지정함. 
기본값은 "blue".}

\item{title}{character. 메인 타이틀.}

\item{subtitle}{character. 서브 타이틀.}

\item{legend_pos}{character. 주제도의 범례를 출력할 위치로 stat 인수를 지정할 
경우에만 효력이 있음. 이 인수를 지정하지 않으면 범례를 출력하지 않음. 
"none", "left", "right", "bottom", "top"에서 선택.}
}
\description{
패키지에 포함된 통계 데이터로 여러 포맷의 주제도를 시각화함
}
\details{
통계의 종류는 shpinfo 데이터 프레임에 32개가 정의되어 있음.
이중 stat_code 변수를 stat 인수에 사용해야 함.
}
\examples{
\dontrun{
# 광역시도 인구분포 주제도 그리기
thematic_map(stat = "population")

# 포인트 추가 및 "Blues" 팔레트 적용
thematic_map(stat = "household", point = TRUE, palette = "Blues")

# 시군구 레벨 서울특별시 가구대비인구수, 범례 추가
thematic_map(zoom = "cty", subset = mega_nm == "서울특별시", stat = "pop_per_hosue", 
             point = TRUE, polygon = FALSE, legend_pos = "bottom")
             
# 초등학교 개수지만 통계량이 출력되지 않음
thematic_map(stat = "elemnt_schl_cnt", point = FALSE, polygon = FALSE)

# 서울 양천구의 평균인구, 이름 라벨 및 범례 추가
thematic_map(zoom = "admi", subset = mega_nm == "서울특별시" & cty_nm \%in\% "양천구", 
             stat = "age_mean", label = "name", legend_pos = "right")
             
# 서울 양천구의 평균인구, 값 라벨 및 범례 제거
thematic_map(zoom = "admi", subset = mega_nm == "서울특별시" & cty_nm \%in\% "양천구", 
             stat = "age_mean", label = "value", legend_pos = "none")
             
# 인구통계의 이름과 값 라벨 추가
thematic_map(stat = "population", zoom = "admi", 
             subset = mega_nm == "서울특별시" & cty_nm \%in\% "양천구", label = "all")
             
# 지역코드로 서울시 시각화
thematic_map(subset = mega_cd == 11)

# 지역코드로 서울시와 경기도 시각화
thematic_map(subset = mega_cd \%in\% c("11", "41"))

# 지역이름으로 수도권 가구수 시각화 
thematic_map(zoom = "cty", stat = "household", 
             subset = mega_nm \%in\% c("서울특별시", "인천광역시", "경기도"))
}
}
\seealso{
\code{\link{mega}}, \code{\link{cty}}, \code{\link{admi}}, 
\code{\link{brewer.pal}}
}
