% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_custom_map}
\alias{theme_custom_map}
\title{map 시각화용 ggplot2 테마}
\usage{
theme_custom_map(
  base_size = 11,
  base_family = "NanumSquare",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  title_size = 16,
  subtitle_size = 12,
  title_margin = 7,
  subtitle_margin = 5,
  ...
)
}
\arguments{
\item{base_size}{numeric. pts로 주어진 기본 글꼴 크기.}

\item{base_family}{character. 기본 글꼴 패밀리. 기본값은 "NanumSquare".}

\item{base_line_size}{numeric. 선 요소의 기본 크기.}

\item{base_rect_size}{numeric. rect 요소의 기본 크기.}

\item{title_size}{numeric. 제목의 폰트 크기.}

\item{subtitle_size}{numeric. 부제목의 폰트 크기.}

\item{title_margin}{numeric. 부제목의 마진 크기.}

\item{subtitle_margin}{numeric. 부제목의 마진 크기.}
}
\description{
플롯팅할 지도를 위한 ggplot2 custom 테마
}
\details{
완전히 비어 있는 테마인 theme_void()를 기반으로 생성
}
\examples{
\donttest{
## 테마 적용 전
mega \%>\% 
  ggplot() +
  geom_sf(aes(fill = household), lwd = 0.3)

## 테마 적용
mega \%>\% 
  ggplot() +
  geom_sf(aes(fill = household), lwd = 0.3) +
  theme_custom_map()
}
}
