% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stats.R
\docType{data}
\name{store_info_seoul}
\alias{store_info_seoul}
\title{상가 위치 정보}
\format{
39개의 변수, 17개 광역시도의 총 레코드 2,532,877을 담은 tbl_df 클래스 객체.
\describe{
  \item{store_id}{character. 상가업소번호.}
  \item{store_nm}{character. 상호명.}
  \item{branch_nm}{character. 지점명.}
  \item{industry_l_cd}{character. 상권업종대분류코드.}
  \item{industry_l_nm}{character. 상권업종대분류명.}
  \item{industry_m_cd}{character. 상권업종중분류코드.}
  \item{industry_m_nm}{character. 상권업종중분류명.}
  \item{industry_s_cd}{character. 상권업종소분류코드.}
  \item{industry_s_nm}{character. 상권업종소분류명.}
  \item{stnd_industry_cd}{character. 표준산업분류코드.}
  \item{stnd_industry_nm}{character. 표준산업분류명.}
  \item{mega_cd}{character. 광역시도 코드.}
  \item{mega_nm}{character. 광역시도 이름.}
  \item{cty_cd}{character. 시군구 코드.}
  \item{cty_nm}{character. 시군구 이름.}
  \item{admi_cd}{character. 읍면동 코드.}
  \item{admi_nm}{character. 읍면동 이름.}
  \item{law_admi_cd}{character. 법정동코드.}
  \item{law_admi_nm}{character. 법정동명.}
  \item{landno_cd}{character. 지번코드.}
  \item{plat_class_cd}{character. 대지구분코드.}
  \item{plat_class_nm}{character. 대지구분명.}
  \item{land_major_no}{integer. 지번본번지.}
  \item{land_minor_no}{integer. 지번부번지.} 
  \item{land_addr}{character. 지번주소.} 
  \item{road_cd}{character. 도로명코드.}  
  \item{road_nm}{character. 도로명.} 
  \item{build_major_no}{integer. 건물본번지.}
  \item{build_minor_no}{integer. 건물부번지.}
  \item{regist_bukd_no}{character. 건물관리번호.}
  \item{build_nm}{character. 건물명.}
  \item{road_addr}{character. 도로명주소.}
  \item{post_cd_old}{character. 구우편번호.}
  \item{post_cd}{character. 신우편번호.}
  \item{dong_info}{character. 동정보.}
  \item{floor_info}{character. 층정보.}   
  \item{house_info}{character. 호정보.}
  \item{lon}{numeric. 위치정보, 경도.}
  \item{lat}{numeric. 위치정보, 위도.}
}
}
\source{
"공공데이터 포털의 소상공인시장진흥공단_상가(상권)정보" in <https://www.data.go.kr/data/15083033/fileData.do>.
}
\usage{
data(store_info_seoul)

data(store_info_gyeonggi)

data(store_info_middle)

data(store_info_south)
}
\description{
소상공인시장진흥공단에서 제공하는 상가 위치 정보
}
\details{
이 데이터셋은 개별 상가의 위치정보 데이터로 상가 건물 위치의 경도/위도 정보를 
포함하고 있음. 
데이터 파일 사이즈가 github에 올릴 수 있는 100MB를 초과하므로 다음처럼 분리함.:
\itemize{
\item store_info_seoul : 서울특별시.
\item store_info_gyeonggi : 경기도.
\item store_info_middle : 중부지방. 인천광역시, 강원도, 세종특별자치시, 
충청남도, 충청북도, 대전광역시.
\item store_info_south : 남부지방. 부산광역시, 대구광역시, 광주광역시, 
울산광역시, 전라북도, 전라남도, 경상북도, 경상남도, 제주특별자치도.
}
}
\keyword{datasets}
